+++ src/main/java/calculator/domain/InputConverter.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package calculator.domain;
 
 import java.util.Arrays;
 import java.util.List;
