+++ src/main/java/calculator/domain/OperatorGroup.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package calculator.domain;
 
 import java.util.ArrayList;
 import java.util.LinkedList;
