+++ src/main/java/calculator/utility/Operator.java
@@ -1,4 +1,4 @@
-package lotto.utility;
+package calculator.utility;
 
 import java.util.Arrays;
 import java.util.function.BinaryOperator;
