+++ src/test/java/calculator/InputConverterTest.java
@@ -1,6 +1,6 @@
-package lotto;
+package calculator;
 
-import lotto.domain.InputConverter;
+import calculator.domain.InputConverter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
